fix: unify typescript, esm import interfaes
Add an esm wrapper so that esm code importing the module sees the same
imports as typescript would, rather than the commonjs imports.

Previously, importing the module in esm returned a different result
than the documentation suggests.

This may arguably be a breaking change. It prevents importation from
subpaths (e.g. 'iterare/filter'), which while not documented to work,
previously likely did. The exposed module is changed, but the
documented portions are not.

Testing this change requires self-referential imports, which do not
work in node <=11 or early versions of 12 or 13. The engine
requirement is now different for development than for consumption,
which is not documentable in package.json at present.

The package.exports feature may be broken in early versions of 12 and
13. Further testing is needed.